refactor: MovieFetcher 추상 객체 적용
